refactor/#87: 파일 관련 로직을 ImageUriManager로 분리
